Report generation; better validation; sizeof counting; XML fixes
Added functionality to count the number of bytes in RAM a struct takes. Verified with SaveBlock1 and the offset annotations in pokeemerald source (the annotations for SaveBlock2 are incorrect), and via Godbolt and sizeof/offsetof checks as well.

The `skip_when_serializing` flag on members should now be honored.

Sector size limits now use named constants instead of magic numbers.

We now generate a full report on the results of serialization, including comparisons for the number of bytes saved.

We no longer generate any code files if any sector groups would overflow their max sector counts.

For some strings in game data, the terminator is actually optional; we now support these.

Fixed various mistakes in the XML.

+++ codegen-main/codegen-main/tests/h/struct-members/BerryPickingResults.members.inl
@@ -1,6 +1,6 @@
    u32 bestScore;
-   u32 berriesPicked;
-   u32 berriesPickedInRow;
+   u16 berriesPicked;
+   u16 berriesPickedInRow;
    u8 field_8;
    u8 field_9;
    u8 field_A;
