+++ codegen-main/codegen-main/tests/h/struct-members/ExternalEventData.members.inl
@@ -7,4 +7,4 @@
    u32 receivedAgetoCelebi : 1; // from JP Colosseum Bonus Disc
    u32 unknownExternalDataFields3 : 4;
    u32 totalEarnedPokeCoupons : 24; // Used by the JP Colosseum bonus disc. Determines PokéCoupon rank to distribute rewards. Unread in International games. Colosseum/XD caps this at 9,999,999.
-   u32 unknownExternalDataFields4[5]; // if actually used, may be broken up into different fields.
+   u8 unknownExternalDataFields4[5]; // if actually used, may be broken up into different fields.
